refactor: Mudando precoEmCentavo de int para long

+++ src/main/java/dev/rafael/DesafioRelogio/entity/Relogio.java
@@ -63,7 +63,7 @@ public class Relogio {
 	private int larguraLugMm;
 
 	@Column(nullable = false)
-	private int precoEmCentavos;
+	private long precoEmCentavos;
 
 	@Column(nullable = false, length = 600)
 	private String urlImagem;
